experiments/requirements.txt
websockets>=12.0
anthropic>=0.40.0
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0
aiofiles>=23.0
cryptography>=42.0
